# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cancun Chaos Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations:
    get:
      operationId: listLocations
      tags: [map]
      summary: List all map locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [map, admin]
      summary: Create a location (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    patch:
      operationId: updateLocation
      tags: [map, admin]
      summary: Update a location (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
    delete:
      operationId: deleteLocation
      tags: [map, admin]
      summary: Delete a location (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /houses/{playerId}:
    get:
      operationId: getHouse
      tags: [map]
      summary: Get player's house
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: House
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
    patch:
      operationId: updateHouse
      tags: [map]
      summary: Update player's house
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHouseBody"
      responses:
        "200":
          description: Updated house
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
  /houses/{playerId}/visit:
    post:
      operationId: visitHouse
      tags: [map, social]
      summary: Visit another player's house
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitHouseBody"
      responses:
        "200":
          description: Visit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitResult"

  # --- EVENTS ---
...
  /minigames:
    get:
      operationId: listMinigames
      tags: [minigames]
      summary: List all minigame definitions
      responses:
        "200":
          description: Minigames
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Minigame"
    post:
      operationId: createMinigame
      tags: [minigames, admin]
      summary: Create a minigame (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMinigameBody"
      responses:
        "201":
          description: Created minigame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Minigame"
  /minigames/{id}:
    patch:
      operationId: updateMinigame
      tags: [minigames, admin]
      summary: Update a minigame (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMinigameBody"
      responses:
        "200":
          description: Updated minigame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Minigame"
  /minigames/{id}/play:
    post:
      operationId: playMinigame
      tags: [minigames]
      summary: Submit minigame result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayMinigameBody"
      responses:
        "200":
          description: Play result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MinigameResult"

  # --- MISSIONS ---
  /missions:
    get:
      operationId: listMissions
      tags: [missions]
      summary: List all missions
      responses:
        "200":
          description: Missions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mission"
    post:
      operationId: createMission
      tags: [missions]
      summary: Create a mission (player or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMissionBody"
      responses:
        "201":
          description: Created mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}:
    patch:
      operationId: updateMission
      tags: [missions, admin]
      summary: Update a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMissionBody"
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
    delete:
      operationId: deleteMission
      tags: [missions, admin]
      summary: Delete a mission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /missions/{id}/complete:
    post:
      operationId: completeMission
      tags: [missions]
      summary: Complete a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteMissionBody"
      responses:
        "200":
          description: Completion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionResult"

  # --- INVENTORY ---
...
components:
  schemas:
...
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        hexQ:
          type: integer
        hexR:
          type: integer
        emoji:
          type: string
        description:
          type: string
        riskLevel:
          type: integer
        energyRestore:
          type: integer
        minigameId:
          type: ["integer", "null"]
      required:
        - id
        - name
        - type
        - hexQ
        - hexR
        - emoji
        - description
        - riskLevel
        - energyRestore
    CreateLocationBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        hexQ:
          type: integer
        hexR:
          type: integer
        emoji:
          type: string
        description:
          type: string
        riskLevel:
          type: integer
        energyRestore:
          type: integer
        minigameId:
          type: ["integer", "null"]
      required:
        - name
        - type
        - hexQ
        - hexR
        - emoji
        - description
        - riskLevel
        - energyRestore
    House:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
        hexQ:
          type: integer
        hexR:
          type: integer
        name:
          type: string
        style:
          type: string
        color:
          type: string
        decoration:
          type: string
        visitCount:
          type: integer
      required:
        - id
        - playerId
        - playerName
        - hexQ
        - hexR
        - name
        - style
        - color
        - decoration
        - visitCount
    UpdateHouseBody:
      type: object
      properties:
        name:
          type: string
        style:
          type: string
        color:
          type: string
        decoration:
          type: string
    VisitHouseBody:
      type: object
      properties:
        visitorPlayerId:
          type: integer
      required:
        - visitorPlayerId
    VisitResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        rewardMoney:
          type: integer
      required:
        - success
        - message
        - rewardMoney
...
    Minigame:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: string
        baseReward:
          type: integer
        difficulty:
          type: integer
        emoji:
          type: string
        config:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - name
        - type
        - description
        - baseReward
        - difficulty
        - emoji
        - config
        - isActive
    CreateMinigameBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        baseReward:
          type: integer
        difficulty:
          type: integer
        emoji:
          type: string
        config:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - type
        - description
        - baseReward
        - difficulty
        - emoji
        - config
        - isActive
    PlayMinigameBody:
      type: object
      properties:
        playerId:
          type: integer
        score:
          type: integer
        accuracy:
          type: number
      required:
        - playerId
        - score
        - accuracy
    MinigameResult:
      type: object
      properties:
        success:
          type: boolean
        score:
          type: integer
        reward:
          type: integer
        message:
          type: string
        newMoney:
          type: integer
      required:
        - success
        - score
        - reward
        - message
        - newMoney
    Mission:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        creatorPlayerId:
          type: ["integer", "null"]
        targetPlayerId:
          type: ["integer", "null"]
        reward:
          type: integer
        type:
          type: string
        status:
          type: string
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - reward
        - type
        - status
        - createdAt
    CreateMissionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        creatorPlayerId:
          type: integer
        targetPlayerId:
          type: ["integer", "null"]
        reward:
          type: integer
        type:
          type: string
      required:
        - title
        - description
        - creatorPlayerId
        - reward
        - type
    UpdateMissionBody:
      type: object
      properties:
        status:
          type: string
        reward:
          type: integer
    CompleteMissionBody:
      type: object
      properties:
        playerId:
          type: integer
      required:
        - playerId
    MissionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        rewardEarned:
          type: integer
      required:
        - success
        - message
        - rewardEarned
...